WinAuth topic enhancements (#10094)
