--- aspnetcore/security/authentication/windowsauth/sample/launchSettings.json
@@ -1,10 +1,0 @@
-{
-    "iisSettings": {
-        "windowsAuthentication": true,
-        "anonymousAuthentication": false,
-        "iisExpress": {
-            "applicationUrl": "http://localhost:52171/",
-            "sslPort": 0
-        }
-    } // additional options trimmed
-}
