+++ aspnetcore/security/authentication/windowsauth/sample_snapshot/Program.cs
