+++ aspnetcore/security/authentication/windowsauth/sample_snapshot/Startup.cs
@@ -1,0 +1,25 @@
+﻿app.Run(async (context) =>
+{
+    try
+    {
+        var user = (WindowsIdentity)context.User.Identity;
+
+        await context.Response
+            .WriteAsync($"User: {user.Name}\tState: {user.ImpersonationLevel}\n");
+
+        WindowsIdentity.RunImpersonated(user.AccessToken, () =>
+        {
+            var impersonatedUser = WindowsIdentity.GetCurrent();
+            var message =
+                $"User: {impersonatedUser.Name}\t" +
+                $"State: {impersonatedUser.ImpersonationLevel}";
+
+            var bytes = Encoding.UTF8.GetBytes(message);
+            context.Response.Body.Write(bytes, 0, bytes.Length);
+        });
+    }
+    catch (Exception e)
+    {
+        await context.Response.WriteAsync(e.ToString());
+    }
+});
