+++ aspnetcore/security/authentication/windowsauth/sample_snapshot/launchSettings.json
@@ -1,0 +1,8 @@
+"iisSettings": {
+    "windowsAuthentication": true,
+    "anonymousAuthentication": false,
+    "iisExpress": {
+        "applicationUrl": "http://localhost:52171/",
+        "sslPort": 0
+    }
+}
