+++ aspnetcore/security/authentication/windowsauth/sample_snapshot/web_1.config
@@ -1,0 +1,8 @@
+<system.webServer>
+  <security>
+    <authentication>
+      <anonymousAuthentication enabled="false" />
+      <windowsAuthentication enabled="true" />
+    </authentication>
+  </security>
+</system.webServer>
