+++ aspnetcore/security/authentication/windowsauth/sample_snapshot/web_2.config
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <location path="." inheritInChildApplications="false">
+    <system.webServer>
+      <security>
+        <authentication>
+          <anonymousAuthentication enabled="false" />
+          <windowsAuthentication enabled="true" />
+          </authentication>
+        </security>
+    </system.webServer>
+  </location>
+</configuration>
